+++ lib/prompts/letter-quality-check.ts
